Updated Testing
Added some print statements to clarify the intervals for increasing and decreasing, the cumulative percent change over each checked series of intervals, and cumulative percent change after owning a crypto for x amount of time. ( MAY BE MORE but they are are all just print statements that provide additional info).