Modified Volume Calculation, Too Low, and Reset Values
I modified the volume calculation so that it attempts to split a volume up based on the % change in that interval. I.e. -20% makes most of the volume count as negative but some will still be counted as positive. I modified Too Low to take the last 10 min of changes into account before selling because it was selling too often. I also modified Too Low to take a value as to how many times over an interval it should increase. I set the RESTART variables to reset at the right time.