New Files and Cleaned Skeleton
Created a Cyrpto Trainer that can run multiple instances of the bot, a Crypto Evaluator (cleaned version of gaybot), and a text file with the best parameters. Crypto Evaluator does not run because the requests are still there and those will need to be changed once we can pull in data with Crypto Trainer and sort through that data. Furthermore, the parameters need to be passed uniquely to each version of CryptoEvaluator. CryptoTrainer must also pull the parameters from the best parameters text file and provide different versions to each 
instance.

Private data just has the api key and the secret key separate from the bot