Check Logs
Changes to cryptoevaluator changed the way modified volumes are calcualted to avoid having 0 as the modified volume for each crypto. Also, cryptoevaluator was modified to ensure that if no crypto is chosen that methods requiring a currency are not run. CryptoTrainer now has the capacity to run multiple variations and then once the best one is chosen a brand new set of variations are run.  While there are no errors that prevent runtime there still could be some errors in reading and writing from the text file. I.e. after it is modified by one set of classes will the new ones read in the proper parameters.