Bug fixes and Intervals
Fixed bug where crypto evaluator always grabbed the same volume by hour and price by hour.
Added in an interval and endpoint parameters to allow for sepcification of the length of a test and when it should end.
Fixed bug where parameters rewriting relied on the exact script of the last parameter