fixed modified Volume and new parameter minOffset
modified volume no longer has issues with the amount of stored values being the same

new parameter min offset is used to tell the bot how much data it should grab before the chosen interval. this data will not be simulated on. instead it will be used as a kind of historical data for the bot