money params, true percent change, new grahs
added a crypto graph script that will draw graphs of data from a passed interval

the true percent change is now recorded.

the two new parameters are used to calculate this, start money is the amount of money the bot is allowed to start with. end money is the calculated end money.