more websockets and distribution work
Fixed websockets using the correct bid or ask based on buying and selling fully separated the two and is now calculating loss and volume correctly and creating the proper list of tradable cryptos at any given time. Got distribution to start pickling the files and can retrieve and add to the lists to create weekly distribution data for every 10 minute interval that we can do analysis on.