SuperTrainer, CryptoTrainer, CryptoEvaluator Working
When run independently the three main bot files will use a different
directory prefix so that they do not write over the code for testing them separately.

Datastream now creates and stores a pricesymbols dictionary and list file whenever it is first run so that the evaluators can operate off of whatever cryptos are actually stored in the database.

All new parameters have been incorporated in crypto evaluator. To add new parameters requires their writing into the Generics files that determine how they are handled (if they are data you must also write a line of code into updateCrypto in CryptoEvaluator to specify what data that will be gathered as, no other changes will be necessary).

CryptoTrainer will train a single evaluator parameter file with classes and variations within a class (think of each teacher as a grade school teacher and the classes are the years they are teaching and the variations are the number of students they have per year). The CryptoTrainer will randomize the normal parameters and add in combined parameters as it sees fit. (so far I have not found evidence of any combined parameters being added).

SuperTrainer grabs a trainer and a stored evaluator file (multiple evaluator files are stored per trainer param file) and then sets up a scenario to train that evaluator with the parameters in its associated trainer param file. SuperTrainer may also be directed to randomize some of the trainer param files before training so that the evaluators are trained differently. SuperTrainer decides whether a trainer and evaluator are better than their stored originals and stores them as it sees fit. SuperTrainer also writes to a file indicating which files have been trained most recently so any higher level can choose.