BETTER VERSION of last update
Noticed last minute that the combined parameters was not adding new parameters as they are needed because none existed. So now new combined parameters will be added and then filled up to a certain limit. There is also a limit given to each trainer that determines how many parameters can be combined per combined parameter.